requirements.txt
requests~=2.27.1
psycopg2~=2.9.3
pandas~=1.4.1
python-dateutil~=2.8.2